ci: fix pipeline for GitHub Actions + resolve all clippy warnings
Backend:
- Remove continue-on-error from Clippy — now enforces -D warnings cleanly
- Fix 27 clippy errors: unused macro, dead structs/fns, redundant closures,
  unnecessary cast, else-if collapse, prefix stripping idiom
- Add #[allow(dead_code)] to test helpers and utility functions kept for
  future use; #![allow(unused_assignments)] for push_set! macro pattern

CI pipeline:
- Add concurrency group to cancel redundant in-progress runs
- Move DATABASE_URL + NEXT_TELEMETRY_DISABLED to workflow-level env
- Remove continue-on-error from Clippy step (now actually enforced)
- Add restore-keys fallback chains for Next.js build cache

+++ backend/src/notifications.rs
@@ -487,8 +487,8 @@ pub async fn get_notifications(
 
     // 14. Sunday backup reminder
     let day_of_week = {
-        let days_since_epoch = (now / 86_400_000) as i64;
-        ((days_since_epoch + 4) % 7)
+        let days_since_epoch = now / 86_400_000;
+        (days_since_epoch + 4) % 7
     }; // 0=Mon..6=Sun
     if day_of_week == 6 && hour >= 10 {
         notifications.push(Notification {
